# openapi.yaml
openapi: 3.0.3
info:
  title: Smart Health Portal API
  version: 1.0.0
  description: >
    Backend for appointments, medical records, and role-based access
    (Patient / Provider). All protected endpoints use Bearer JWT.
servers:
  - url: http://localhost:3000
    description: Local
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MedicalRecord:
      type: object
      properties:
        record_id: { type: integer }
        patient_id: { type: integer }
        file_path: { type: string, format: uri }
        uploaded_at: { type: string, format: date-time }
        file_name: { type: string }
paths:
...
  /api/doctor/patient/{patientId}/records:
    get:
      tags: [Doctors]
      summary: Medical records of this patient (doctor must be associated via an appointment)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: patientId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MedicalRecord' }
  /api/patient/records:
    get:
      tags: [Patients]
      summary: List own medical records
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MedicalRecord' }
  /api/records/upload:
    post:
      tags: [Records]
      summary: Upload a medical file
      description: >
        Patients upload for themselves.
        Providers can upload for a patient by passing `patientId` (form field).
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                patientId:
                  type: integer
                  nullable: true
      responses:
        '200': { description: Uploaded }
  /api/records/{recordId}:
    delete:
      tags: [Records]
      summary: Delete a medical record
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: recordId
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Deleted }
        '403': { description: Not allowed }